+++ src/main/java/de/samply/converter/Format.java
@@ -10,24 +10,26 @@ import java.util.concurrent.atomic.AtomicBoolean;
 
 public enum Format {
 
-    FHIR_SEARCH(String.class, true),
-    CQL(String.class, true),
-    CQL_DATA(String.class, true),
-    MEASURE_REPORT(MeasureReport.class, false),
-    BUNDLE(Bundle.class, false),
-    CONTAINERS(Containers.class, false),
-    CSV(Path.class, false),
-    EXCEL(Path.class, false),
-    OPAL(Path.class, false),
-    JSON(Path.class, false),
-    XML(Path.class, false);
+    FHIR_SEARCH(String.class, true, false),
+    CQL(String.class, true, false),
+    CQL_DATA(String.class, true, false),
+    MEASURE_REPORT(MeasureReport.class, false, false),
+    BUNDLE(Bundle.class, false, false),
+    CONTAINERS(Containers.class, false, false),
+    CSV(Path.class, false, true),
+    EXCEL(Path.class, false, true),
+    OPAL(Path.class, false, true),
+    JSON(Path.class, false, true),
+    XML(Path.class, false, true);
 
     private Class zClass;
     private boolean isQuery;
+    private boolean isOutput;
 
-    Format(Class zClass, boolean isQuery) {
+    Format(Class zClass, boolean isQuery, boolean isOutput) {
         this.zClass = zClass;
         this.isQuery = isQuery;
+        this.isOutput = isOutput;
     }
 
     public boolean isInstance(Object object) {
@@ -48,6 +50,10 @@ public enum Format {
         return isQuery;
     }
 
+    public boolean isOutput() {
+        return isOutput;
+    }
+
     public boolean isPath() {
         return zClass == Path.class;
     }
@@ -58,6 +64,10 @@ public enum Format {
 
     public static String[] fetchNoQueries() {
         return Arrays.stream(values()).filter(format -> !format.isQuery()).map(Enum::name).toArray(String[]::new);
+    }
+
+    public static String[] fetchOutputs() {
+        return Arrays.stream(values()).filter(Format::isOutput).map(Enum::name).toArray(String[]::new);
     }
 
 }